Scrape list of court cases from Curia

Harvests data from lists of EU court cases from curia.europa.eu. CELEX identifiers are extracted from hyperlinks where available.

Usage

elx_curia_list(
  data = c("all", "ecj_old", "ecj_new", "gc_all", "cst_all"),
  parse = TRUE
)

Arguments

data

Data to be scraped from four separate lists of cases maintained by Curia, defaults to "all" which contains cases from Court of Justice, General Court and Civil Service Tribunal.

parse

If TRUE, references to cases and appeals are parsed out from case_info into separate columns

Value

A data frame containing case identifiers and information as character columns. Where the case id contains a hyperlink to Eur-Lex, the CELEX identifier is retrieved as well. Hyperlinks to Eur-Lex disappeared from more recent cases.

Examples

# \donttest{
elx_curia_list(data = "cst_all", parse = FALSE)
#> # A tibble: 1,762 × 3
#>    case_id    case_id_celex case_info                                           
#>    <chr>      <chr>         <chr>                                               
#>  1 F-1/05 *   NA            Judgment of 26 October 2006, Landgren / ETF (F-1/05…
#>  2 F-1/05     NA            Order of 22 May 2007, Landgren / ETF (F-1/05, ECR-S…
#>  3 F-1/05 INT NA            Order of 13 July 2007, Landgren / ETF (F-1/05 INT, …
#>  4 F-1/05     NA            Order of 9 November 2010, Landgren / ETF (F-1/05, u…
#>  5 F-2/05     NA            Removed from the register on 18 June 2008, Kröppeli…
#>  6 F-3/05     NA            Order of 15 May 2006, Schmit / Commission (F-3/05, …
#>  7 F-4/05     NA            Removed from the register on 18 June 2008, Huober /…
#>  8 F-5/05 *   NA            Judgment of 28 April 2009, Violetti and others / Co…
#>  9 F-6/05     NA            Removed from the register on 18 June 2008, Kröppeli…
#> 10 F-7/05     NA            Schmit / Commission (F-7/05) , see Case  F-5/05     
#> # ℹ 1,752 more rows
# }