Scrape list of court cases from Curia

Harvests data from lists of EU court cases from curia.europa.eu. CELEX identifiers are extracted from hyperlinks where available.

elx_curia_list(data = c("all", "ecj_old", "ecj_new", "gc_all", "cst_all"))

Arguments

data
Data to be scraped from four separate lists of cases maintained by Curia, defaults to "all" which contains cases from Court of Justice, General Court and Civil Service Tribunal.

Value

A data frame containing case identifiers and information as character columns.

Examples

# \donttest{
elx_curia_list(data = "cst_all")

#> # A tibble: 1,759 x 3
#>    case_id   case_info                                              linked_celex
#>    <chr>     <chr>                                                  <chr>       
#>  1 F-1/05 *  Judgment of 26 October 2006, Landgren / ETF (F-1/05, ~ NA          
#>  2 F-1/05    Order of 22 May 2007, Landgren / ETF (F-1/05, ECR-SC ~ NA          
#>  3 F-1/05 I~ Order of 13 July 2007, Landgren / ETF (F-1/05 INT, EC~ NA          
#>  4 F-1/05    Order of 9 November 2010, Landgren / ETF (F-1/05, unp~ NA          
#>  5 F-2/05    Removed from the register on 18 June 2008, Kröppelin ~ NA          
#>  6 F-3/05    Order of 15 May 2006, Schmit / Commission (F-3/05, EC~ NA          
#>  7 F-4/05    Removed from the register on 18 June 2008, Huober / C~ NA          
#>  8 F-5/05 *  Judgment of 28 April 2009, Violetti and others / Comm~ NA          
#>  9 F-6/05    Removed from the register on 18 June 2008, Kröppelin ~ NA          
#> 10 F-7/05    Schmit / Commission (F-7/05) , see Case  F-5/05        NA          
#> # ... with 1,749 more rows

# }