Create SPARQL queries

Generates pre-defined or manual SPARQL queries to retrieve document ids from Cellar. List of available resource types: http://publications.europa.eu/resource/authority/resource-type . Note that not all resource types are compatible with default parameter values.

elx_make_query(
  resource_type = c("any", "directive", "regulation", "decision", "recommendation",
    "intagr", "caselaw", "manual", "proposal", "national_impl"),
  manual_type = "",
  directory = NULL,
  sector = NULL,
  include_corrigenda = FALSE,
  include_celex = TRUE,
  include_lbs = FALSE,
  include_date = FALSE,
  include_date_force = FALSE,
  include_date_endvalid = FALSE,
  include_date_transpos = FALSE,
  include_date_lodged = FALSE,
  include_force = FALSE,
  include_eurovoc = FALSE,
  include_author = FALSE,
  include_citations = FALSE,
  include_court_procedure = FALSE,
  include_directory = FALSE,
  include_sector = FALSE,
  include_ecli = FALSE,
  order = FALSE,
  limit = NULL
)

Arguments

resource_type

Type of resource to be retrieved via SPARQL query

manual_type

Define manually the type of resource to be retrieved

directory

Restrict the results to a given directory code

sector

Restrict the results to a given sector code

include_corrigenda

If `TRUE`, results include corrigenda

include_celex

If `TRUE`, results include CELEX identifier for each resource URI

include_lbs

If `TRUE`, results include legal bases of legislation

include_date

If `TRUE`, results include document date

include_date_force

If `TRUE`, results include date of entry into force

include_date_endvalid

If `TRUE`, results include date of end of validity

include_date_transpos

If `TRUE`, results include date of transposition deadline for directives

include_date_lodged

If `TRUE`, results include date a court case was lodged with the court

include_force

If `TRUE`, results include whether legislation is in force

include_eurovoc

If `TRUE`, results include EuroVoc descriptors of subject matter

include_author

If `TRUE`, results include document author(s)

include_citations

If `TRUE`, results include citations (CELEX-labelled)

include_court_procedure

If `TRUE`, results include type of court procedure and outcome

include_directory

If `TRUE`, results include the Eur-Lex directory code

include_sector

If `TRUE`, results include the Eur-Lex sector code

include_ecli

If `TRUE`, results include the ECLI identifier for court documents

order

Order results by ids

limit

Limit the number of results, for testing purposes mainly

Value

A character string containing the SPARQL query

Examples

elx_make_query(resource_type = "directive", include_date = TRUE, include_force = TRUE)
#> [1] "PREFIX cdm: <http://publications.europa.eu/ontology/cdm#>\n  PREFIX annot: <http://publications.europa.eu/ontology/annotation#>\n  PREFIX skos:<http://www.w3.org/2004/02/skos/core#>\n  PREFIX dc:<http://purl.org/dc/elements/1.1/>\n  PREFIX xsd:<http://www.w3.org/2001/XMLSchema#>\n  PREFIX rdf:<http://www.w3.org/1999/02/22-rdf-syntax-ns#>\n  PREFIX owl:<http://www.w3.org/2002/07/owl#>\n  select distinct ?work ?type ?celex str(?date) ?force where{ ?work cdm:work_has_resource-type ?type. FILTER(?type=<http://publications.europa.eu/resource/authority/resource-type/DIR>||\n  ?type=<http://publications.europa.eu/resource/authority/resource-type/DIR_IMPL>||\n  ?type=<http://publications.europa.eu/resource/authority/resource-type/DIR_DEL>) \n FILTER not exists{?work cdm:work_has_resource-type <http://publications.europa.eu/resource/authority/resource-type/CORRIGENDUM>} OPTIONAL{?work cdm:resource_legal_id_celex ?celex.} OPTIONAL{?work cdm:work_date_document ?date.} OPTIONAL{?work cdm:resource_legal_in-force ?force.} }"
elx_make_query(resource_type = "regulation", include_corrigenda = TRUE, order = TRUE)
#> [1] "PREFIX cdm: <http://publications.europa.eu/ontology/cdm#>\n  PREFIX annot: <http://publications.europa.eu/ontology/annotation#>\n  PREFIX skos:<http://www.w3.org/2004/02/skos/core#>\n  PREFIX dc:<http://purl.org/dc/elements/1.1/>\n  PREFIX xsd:<http://www.w3.org/2001/XMLSchema#>\n  PREFIX rdf:<http://www.w3.org/1999/02/22-rdf-syntax-ns#>\n  PREFIX owl:<http://www.w3.org/2002/07/owl#>\n  select distinct ?work ?type ?celex where{ ?work cdm:work_has_resource-type ?type. FILTER(?type=<http://publications.europa.eu/resource/authority/resource-type/REG>||\n  ?type=<http://publications.europa.eu/resource/authority/resource-type/REG_IMPL>||\n  ?type=<http://publications.europa.eu/resource/authority/resource-type/REG_FINANC>||\n  ?type=<http://publications.europa.eu/resource/authority/resource-type/REG_DEL>) OPTIONAL{?work cdm:resource_legal_id_celex ?celex.} } order by str(?date)"
elx_make_query(resource_type = "caselaw")
#> [1] "PREFIX cdm: <http://publications.europa.eu/ontology/cdm#>\n  PREFIX annot: <http://publications.europa.eu/ontology/annotation#>\n  PREFIX skos:<http://www.w3.org/2004/02/skos/core#>\n  PREFIX dc:<http://purl.org/dc/elements/1.1/>\n  PREFIX xsd:<http://www.w3.org/2001/XMLSchema#>\n  PREFIX rdf:<http://www.w3.org/1999/02/22-rdf-syntax-ns#>\n  PREFIX owl:<http://www.w3.org/2002/07/owl#>\n  select distinct ?work ?type ?celex where{ ?work cdm:work_has_resource-type ?type. FILTER(?type=<http://publications.europa.eu/resource/authority/resource-type/JUDG>||\n  ?type=<http://publications.europa.eu/resource/authority/resource-type/ORDER>||\n  ?type=<http://publications.europa.eu/resource/authority/resource-type/OPIN_JUR>||\n  ?type=<http://publications.europa.eu/resource/authority/resource-type/THIRDPARTY_PROCEED>||\n  ?type=<http://publications.europa.eu/resource/authority/resource-type/GARNISHEE_ORDER>||\n  ?type=<http://publications.europa.eu/resource/authority/resource-type/RULING>||\n  ?type=<http://publications.europa.eu/resource/authority/resource-type/JUDG_EXTRACT>||\n  ?type=<http://publications.europa.eu/resource/authority/resource-type/INFO_JUDICIAL>||\n  ?type=<http://publications.europa.eu/resource/authority/resource-type/VIEW_AG>||\n  ?type=<http://publications.europa.eu/resource/authority/resource-type/OPIN_AG>) OPTIONAL{?work cdm:resource_legal_id_celex ?celex.} }"
elx_make_query(resource_type = "manual", manual_type = "SWD")
#> [1] "PREFIX cdm: <http://publications.europa.eu/ontology/cdm#>\n  PREFIX annot: <http://publications.europa.eu/ontology/annotation#>\n  PREFIX skos:<http://www.w3.org/2004/02/skos/core#>\n  PREFIX dc:<http://purl.org/dc/elements/1.1/>\n  PREFIX xsd:<http://www.w3.org/2001/XMLSchema#>\n  PREFIX rdf:<http://www.w3.org/1999/02/22-rdf-syntax-ns#>\n  PREFIX owl:<http://www.w3.org/2002/07/owl#>\n  select distinct ?work ?type ?celex where{ ?work cdm:work_has_resource-type ?type.FILTER(?type=<http://publications.europa.eu/resource/authority/resource-type/SWD>) \n FILTER not exists{?work cdm:work_has_resource-type <http://publications.europa.eu/resource/authority/resource-type/CORRIGENDUM>} OPTIONAL{?work cdm:resource_legal_id_celex ?celex.} }"